requirements.txt
decorator==5.1.1
Django==4.1.0
django-cors-headers==3.14.0
django-pandas==0.6.6
djangorestframework>=3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg>=1.21.4
grpcio==1.51.1
grpcio-reflection==1.51.1
grpcio-tools==1.51.1
icecream==2.1.3
ipdb==0.13.13
ipython==8.11.0
jwt==1.3.1
whitenoise==6.4.0
